backend/requirements.txt
uvicorn==0.34.2
fastapi==0.115.12
pydantic-settings==2.9.1
langchain[openai]==0.3.24
langsmith==0.3.33
python-dotenv==1.1.0
prometheus-fastapi-instrumentator==7.0.0